feat(cli): use Buzz Desktop identity
Ai-assisted: true
