+++ crates/buzz-cli/Cargo.toml
@@ -44,6 +44,7 @@ chrono = { workspace = true }
 # Typed event builders for all write operations
 buzz-sdk = { workspace = true }
 buzz-core = { workspace = true }
+buzz-secret-store = { workspace = true }
 
 # Base64 encoding — NIP-98 event serialization for Authorization header
 base64 = "0.22"
@@ -85,6 +86,9 @@ rustls = { version = "0.23", default-features = false, features = ["ring", "std"
 
 # Random number generation — full jitter for exponential backoff in with_retry
 rand = { workspace = true }
+
+# Clear transient private-key strings after parsing.
+zeroize = { workspace = true }
 
 [dev-dependencies]
 # Scratch files for channel-templates.json fixtures in tests
