+++ desktop/src-tauri/src/managed_agents/storage.rs
@@ -481,7 +481,8 @@ pub fn migrate_agent_keys_to_dev_service(app: &tauri::AppHandle) {
     // A fresh non-singleton store for the prod service — its own empty
     // cache so reads go to the OS keyring without polluting the dev
     // singleton's cache.
-    let prod_store = crate::secret_store::SecretStore::keyring("buzz-desktop");
+    let prod_store =
+        crate::secret_store::SecretStore::keyring(buzz_secret_store::DESKTOP_RELEASE_SERVICE);
     let dev_store = crate::secret_store::SecretStore::shared(keyring_service());
     copy_agent_keys_between_stores(&pubkeys, &prod_store, dev_store);
 }
